Move the kernel into high memory
The kernel is now executed at the address 0xffff ffff fc000 0000.
The kernel used virtual memory and identity map physical memory (or at
least the first 2 Mo). For this reason, the allocation is for the
moment limited  up to 2 Mo (even if the threshold is put at 64 Mo).

The declaration and initialization of the common processor structures
(page tables, IDT, GDT, TSS) is now done in C++ files instead of linker
files. They are now in correct places (in the available, unused RAM).

Finally some files have changed names to improve consistency.