Add user space program + kernel mapping
It is now possible to compile user program using the standard
librairies. The program test1 is mapped in kernel memory for later
testing.

The standard librairi rely on the default Linux syscall. The only
implemented at the moment is the exit (sbrk is not implemented yet). The
librairies use weak symbol for glue functions (exit, sbrk) that are
overwritten by Linux syscalls (user directory), or by a kernel function
(kernel directory).