Parse elf64 file and user virtual memory mapping
The kernel contains the copy of some user programs. The KernelFs provide
an access to them.
The Elf64 parser create a memory mapping (VirtualArea for code, data,
stack, heap, entry point and page table for user space). This mapping can then
be directly used by a processus.

However the suppression of the structure is not taken care of yet. So
the use of the new struct will surely lead to memory leak inside the
kernel.