Time (sleep) and static constructors
The kernel was not yet executing the constructors for static objects.
That is now done. However, in the future, it will be necessary to add
this functionality inside the c library and use the code also for user
apps (which is not yet the case).

The static constructors ware needed to implement the waiting for the
processes:
The processes waiting for an event (here only a timer) are put in the
events list (let's find a better name later). A process call nanosleep
to stop for a few second. it create a timer register to the clock. The
process is linked to the event. Once the timer comes to an end, the
clock ask the processscheduler to wake up (put inside the ready queue)
all the processes waiting on the specific timer.

In the future, other type of event could be add (waiting for an
operation on a file for example). But, be careful, the current
implementation doesn't support processes waiting on multiple events
(like for a poll on multiple file descriptors).

The data structures used are also far from optimal, a priority queue
would be better than the timer list in the clock and a bidirectional
hashmap would be better than the sleeping process list.